requirements.txt
fastapi
uvicorn
numpy
torch
torchaudio
pydantic
omegaconf
pytest
python-multipart
ffmpeg-python
python-dotenv
requests
websockets
soundfile
librosa
PyYAML
nemo_toolkit[asr]
uvloop
httptools
cuda-python